+++ include/Manager/vpg_cpp_generation_manager.hpp
@@ -11,7 +11,7 @@ using namespace vcc;
 class VPGCppGenerationManager : public VPGBaseGenerationManager<VPGCppGenerationManager>
 {
     public:
-        VPGCppGenerationManager(std::shared_ptr<LogProperty> logProperty, std::shared_ptr<VPGGenerationOption> option) : VPGBaseGenerationManager(logProperty, option) {}
+        VPGCppGenerationManager(std::shared_ptr<LogProperty> logProperty, std::wstring workspace, std::shared_ptr<VPGGenerationOption> option) : VPGBaseGenerationManager(logProperty, workspace, option) {}
         virtual ~VPGCppGenerationManager() {}
 
         virtual void Add() const override;